# perf: faster JumpTable bits lookup