--- a/crates/bytecode/src/legacy/analyzed.rs
+++ b/crates/bytecode/src/legacy/analyzed.rs
@@ -60,11 +60,10 @@ impl LegacyAnalyzedBytecode {
         if original_len > bytecode.len() {
             panic!("original_len is greater than bytecode length");
         }
-        if original_len > jump_table.0.len() {
+        if original_len > jump_table.len {
             panic!(
                 "jump table length {} is less than original length {}",
-                jump_table.0.len(),
-                original_len
+                jump_table.len, original_len
             );
         }
 
@@ -116,7 +115,6 @@ mod tests {
     use super::*;
     use crate::{opcode, LegacyRawBytecode};
     use bitvec::{bitvec, order::Lsb0};
-    use std::sync::Arc;
 
     #[test]
     fn test_bytecode_new() {
@@ -142,7 +140,7 @@ mod tests {
     fn test_panic_on_short_jump_table() {
         let bytecode = Bytes::from_static(&[opcode::PUSH1, 0x01]);
         let bytecode = LegacyRawBytecode(bytecode).into_analyzed();
-        let jump_table = JumpTable(Arc::new(bitvec![u8, Lsb0; 0; 1]));
+        let jump_table = JumpTable::new(bitvec![u8, Lsb0; 0; 1]);
         let _ = LegacyAnalyzedBytecode::new(bytecode.bytecode, bytecode.original_len, jump_table);
     }
 
@@ -150,7 +148,7 @@ mod tests {
     #[should_panic(expected = "last bytecode byte should be STOP (0x00)")]
     fn test_panic_on_non_stop_bytecode() {
         let bytecode = Bytes::from_static(&[opcode::PUSH1, 0x01]);
-        let jump_table = JumpTable(Arc::new(bitvec![u8, Lsb0; 0; 2]));
+        let jump_table = JumpTable::new(bitvec![u8, Lsb0; 0; 2]);
         let _ = LegacyAnalyzedBytecode::new(bytecode, 2, jump_table);
     }
 
@@ -158,7 +156,7 @@ mod tests {
     #[should_panic(expected = "bytecode cannot be empty")]
     fn test_panic_on_empty_bytecode() {
         let bytecode = Bytes::from_static(&[]);
-        let jump_table = JumpTable(Arc::new(bitvec![u8, Lsb0; 0; 0]));
+        let jump_table = JumpTable::new(bitvec![u8, Lsb0; 0; 0]);
         let _ = LegacyAnalyzedBytecode::new(bytecode, 0, jump_table);
     }
 }