Cache `Arc<BitVec<u8>>` raw pointer and use bit operations for 7-8x performance improvement in `JumpTable` access/lookups via `is_valid`. Fixes https://github.com/bluealloy/revm/issues/2437

## Benchmark

Implementation    | Time per op | Ops/sec | Speedup
------------------|-------------|---------|--------
Cached Pointer    | 13.40 ns    | 74.6M   | 7.9x faster
Arc Deref         | 106.26 ns   | 9.4M    | baseline

Pattern     | Cached    | Arc       | Speedup
------------|-----------|-----------|--------
Sequential  | 16.44 ns  | 100.77 ns | 6.1x
Random      | 12.57 ns  | 98.48 ns  | 7.8x

Both sequential and random access patterns show 6-8x speedup. All tests pass with identical results

PS: I included the benchmark test so anyone can run see the result <3